apps/voice-agent/requirements.txt
# LiveKit Agents Framework
livekit-agents>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-deepgram>=0.8.0
livekit-plugins-silero>=0.8.0

# API Communication (async, non-blocking)
httpx>=0.25.0
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
structlog>=23.0.0
